Move OpenGraph code into its own file
